requirements.txt
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cmake==3.25.0
customtkinter==5.0.3
darkdetect==0.7.1
et-xmlfile==1.1.0
google==3.0.0
google-api-core==1.32.0
google-api-python-client==1.8.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
html2image==2.0.1
httplib2==0.20.4
idna==3.3
numpy==1.23.2
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySimpleGUI==4.60.3
PySimpleGUIDesigner==0.1.4.6
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
uritemplate==3.0.1
urllib3==1.26.11
